Revert "Specular light WIP"
This reverts commit a28d48c04dca997ff56f9e7db455dcf15ade1516.

+++ src/headers/Light.h
@@ -9,7 +9,7 @@ public:
     Light();
     Light(GLfloat red, GLfloat green, GLfloat blue, GLfloat ambientIntensity, GLfloat xDir, GLfloat yDir, GLfloat zDir, GLfloat diffuseIntensity);
 
-    void UseLight(GLuint ambientIntesityLocation, GLuint ambienColourLocation, GLuint diffuseIntensityLocation, GLuint directionLocation);
+    void UseLight(GLfloat ambientIntesityLocation, GLfloat ambienColourLocation, GLfloat diffuseIntensityLocation, GLfloat directionLocation);
 
     ~Light();
